move globals to global_state file
